Add some module aliases to fix melange builds
There is some issue currently with melange/dune integration, where ocamldep won't see modules referenced in values that are behind Melange ppx, like `bs.obj`.

In this case, `bs.obj` is used to create the `config` value, and `Crypto` and `NodeLoader` are referenced from it. Adding an alias fixes the issue so that ocamldep can see the usage and build the referenced modules.

See https://github.com/ocaml/dune/pull/6625 for more details.